requirements.txt
rootutils
paho-mqtt
python-weather
suntime


